requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
spacy==3.7.4
scikit-learn==1.5.0
matplotlib==3.9.0
plotly==5.22.0
torch==2.3.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
joblib==1.4.2
